#8: Implement checkLIC6

+++ src/test/java/org/example/CMVCalculatorTest.java
@@ -8,9 +8,9 @@ public class CMVCalculatorTest {
 
     @Test
     public void testBasicTwoPointsLIC0() {
-        float[] point1 = new float[]{1.0f, 2.0f};
-        float[] point2 = new float[]{1.0f, -5.0f};
-        float[][] points = new float[][]{point1, point2};
+        double[] point1 = new double[]{1.0f, 2.0f};
+        double[] point2 = new double[]{1.0f, -5.0f};
+        double[][] points = new double[][]{point1, point2};
         assertTrue(CMVCalculator.checkLIC0(points, 6.0d));
         assertFalse(CMVCalculator.checkLIC0(points, 10.0d));
         assertFalse(CMVCalculator.checkLIC0(points, -1.0d));
